requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
Flask-Mail
python-dotenv
requests
